msm: kgsl: Use lock-less list for page pools
Page pool additions and removals are very hot during GPU workloads, so
they should be optimized accordingly. We can use a lock-less list for
storing the free pages in order to speed things up. The lock-less list
allows for one llist_del_first() user and unlimited llist_add() users to
run concurrently, so only a spin lock around the llist_del_first() is
needed; everything else is lock-free. The per-pool page count is now an
atomic to make it lock-free as well.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I51f1f5831663b431b28c3bf3a613619928e27835
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>

 Conflicts:
	drivers/gpu/msm/kgsl_pool.c